+++ frontend/src/components/Topology/css/topology-node.css
@@ -13,6 +13,13 @@
     fill: #394b54;
     cursor: default;
 }
+.node text.resourceCountText {
+    font-size: 16px;
+    font-weight: bold;
+    fill: black;
+    stroke: none;
+}
+
 .node .title {
     font-weight: bold;
 }
