+++ frontend/src/components/Topology/shapes/DiagramShapes.tsx
@@ -1,8 +1,15 @@
 /* Copyright Contributors to the Open Cluster Management project */
 export const DiagramShapes = () => (
     <svg xmlns="http://www.w3.org/2000/svg" xmlnsXlink="http://www.w3.org/1999/xlink" width="0" height="0">
-        <symbol viewBox="0 0 41.5 41.3" id="diagramShapes_circle">
-            <circle cx="20.8" cy="20.7" r="19.1" />
+        <symbol viewBox="0 0 36 36" id="diagramShapes_select">
+            <circle cx="18" cy="18" r="16" strokeWidth="2" />
+        </symbol>
+        <symbol viewBox="0 0 64 46" id="diagramIcons_selectMultiplier">
+            <rect x="2" y="2" width="60px" height="44px" rx="22px" ry="22px" />
+        </symbol>
+        <symbol viewBox="0 0 32 32" id="diagramShapes_multiplier">
+            <circle cx="25.795" cy="15.795" r="15.093" fill="#F0F0F0" stroke="#c7c7c7" strokeWidth="1" />
+            <circle cx="20.795" cy="15.795" r="15.093" fill="#F0F0F0" stroke="#c7c7c7" strokeWidth="1" />
         </symbol>
         <symbol viewBox="0 0 32 32" id="diagramShapes_application">
             <g id="Application" fill="none" fillRule="evenodd">
@@ -223,7 +230,7 @@ export const DiagramShapes = () => (
                     <g id="Namespace" transform="translate(1245.000000, 1091.000000)">
                         <circle
                             id="Oval-Copy-38"
-                            stroke="#82CFFF"
+                            stroke="#c7c7c7"
                             strokeWidth="1.404"
                             fill="#FFFFFF"
                             cx="15.795"
@@ -287,7 +294,7 @@ export const DiagramShapes = () => (
         </symbol>
         <symbol viewBox="0 0 32 32" id="diagramShapes_placement">
             <g id="Placement" transform="translate(0)" fillRule="evenodd">
-                <circle cx="15.795" cy="15.795" r="15.093" fill="#fff" stroke="#3ddbd9" strokeWidth="1.404" />
+                <circle cx="15.795" cy="15.795" r="15.093" fill="#fff" stroke="#c7c7c7" strokeWidth="1.404" />
                 <g transform="matrix(1.8606 0 0 1.8606 6.8571 6.3571)" stroke="none" fill="#000000">
                     <path
                         id="Fill"
