+++ frontend/src/routes/Applications/ApplicationDetails/ApplicationTopology/model/topologySubscription.js
@@ -1,6 +1,7 @@
 /* Copyright Contributors to the Open Cluster Management project */
 
-import { get, includes, concat, uniqBy, filter, keyBy, cloneDeep } from 'lodash'
+import { get, includes, concat, uniqBy, filter, keyBy, cloneDeep, groupBy } from 'lodash'
+
 import { createChildNode, addClusters } from './utils'
 
 const localClusterName = 'local-cluster'
@@ -16,7 +17,7 @@ export const getSubscriptionTopology = (application, managedClusters, relatedRes
     const allAppClusters = application.allClusters ? application.allClusters : []
     const appId = `application--${name}`
     nodes.push({
-        name,
+        name: '',
         namespace,
         type: 'application',
         id: appId,
@@ -41,10 +42,8 @@ export const getSubscriptionTopology = (application, managedClusters, relatedRes
         return cluster?.metadata?.name
     })
     // if application has subscriptions
-    let parentId
     let clusterId
     if (application.subscriptions) {
-        const createdClusterElements = new Set()
         application.subscriptions.forEach((subscription) => {
             // get cluster placement if any
             const ruleDecisionMap = {}
@@ -73,49 +72,42 @@ export const getSubscriptionTopology = (application, managedClusters, relatedRes
                 ruleDecisionMap[localClusterName] = localClusterName
             }
             const ruleClusterNames = Object.keys(ruleDecisionMap)
+
+            // get source
+            const ann = get(subscription, 'metadata.annotations', {})
+            let source =
+                ann['apps.open-cluster-management.io/git-path'] ||
+                ann['apps.open-cluster-management.io/github-path'] ||
+                ann['apps.open-cluster-management.io/bucket-path'] ||
+                get(subscription, 'spec.packageOverrides[0].packageName') ||
+                ''
+            source = source.split('/').pop()
+
+            // add cluster nodes
+            clusterId = addClusters(appId, subscription, source, ruleClusterNames, managedClusterNames, links, nodes)
+
             const isRulePlaced = ruleClusterNames.length > 0
 
             // add subscription node
-            parentId = addSubscription(appId, subscription, isRulePlaced, links, nodes)
+            const subscriptionId = addSubscription(clusterId, subscription, isRulePlaced, links, nodes)
 
             // add rules node
             if (subscription.rules) {
-                addSubscriptionRules(parentId, subscription, links, nodes)
+                addSubscriptionRules(clusterId, subscription, links, nodes)
             }
 
             // add prehooks
             if (subscription.prehooks && subscription.prehooks.length > 0) {
-                addSubscriptionHooks(parentId, subscription, links, nodes, true)
+                addSubscriptionHooks(subscriptionId, subscription, links, nodes, true)
             }
             if (subscription.posthooks && subscription.posthooks.length > 0) {
-                addSubscriptionHooks(parentId, subscription, links, nodes, false)
+                addSubscriptionHooks(subscriptionId, subscription, links, nodes, false)
             }
 
-            // if no cluster found by the placement, use a default empty cluster name so that the deployables are parsed and shown
-            let clusterShapes = [['']]
-            if (ruleClusterNames.length > 1) {
-                clusterShapes = [ruleClusterNames]
-            } else if (ruleClusterNames.length === 1) {
-                clusterShapes = ruleClusterNames.map((cn) => [cn])
+            // add deployed resource nodes using subscription report
+            if (subscription.report) {
+                processReport(subscription.report, subscriptionId, links, nodes, relatedResources)
             }
-
-            // add cluster nodes
-            clusterShapes.forEach((clusterNames) => {
-                clusterId = addClusters(
-                    parentId,
-                    createdClusterElements,
-                    subscription,
-                    clusterNames,
-                    managedClusterNames,
-                    links,
-                    nodes
-                )
-
-                // add deployed resource nodes using subscription report
-                if (subscription.report) {
-                    processReport(subscription.report, clusterId, links, nodes, relatedResources)
-                }
-            })
         })
     }
 
@@ -128,7 +120,6 @@ const addSubscription = (appId, subscription, isPlaced, links, nodes) => {
     } = subscription
     const subscriptionId = `member--subscription--${namespace}--${name}`
     const rule = get(subscription, 'rules[0]')
-
     nodes.push({
         name,
         namespace,
@@ -206,14 +197,15 @@ const addSubscriptionHooks = (parentId, subscription, links, nodes, isPreHook) =
 const processReport = (report, clusterId, links, nodes, relatedResources) => {
     // for each resource, add what it's related to
     report = cloneDeep(report)
+    const resources = report.resources || []
     if (relatedResources) {
-        report.resources.forEach((resource) => {
+        resources.forEach((resource) => {
             const { name, namespace } = resource
             resource.template = relatedResources[`${name}-${namespace}`]
         })
     }
 
-    const serviceOwners = filter(report.resources, (obj) => {
+    const serviceOwners = filter(resources, (obj) => {
         const kind = get(obj, 'kind', '')
         return includes(['Route', 'Ingress', 'StatefulSet'], kind)
     })
@@ -221,7 +213,7 @@ const processReport = (report, clusterId, links, nodes, relatedResources) => {
     // process route and service first
     const serviceMap = processServiceOwner(clusterId, serviceOwners, links, nodes, relatedResources)
 
-    const services = filter(report.resources, (obj) => {
+    const services = filter(resources, (obj) => {
         const kind = get(obj, 'kind', '')
         return includes(['Service'], kind)
     })
@@ -230,14 +222,30 @@ const processReport = (report, clusterId, links, nodes, relatedResources) => {
     processServices(clusterId, services, links, nodes, serviceMap)
 
     // then the rest
-    const other = filter(report.resources, (obj) => {
+    const others = filter(resources, (obj) => {
         const kind = get(obj, 'kind', '')
         return !includes(['Route', 'Ingress', 'StatefulSet', 'Service'], kind)
     })
 
-    other.forEach((resource) => {
+    processMultiples(others).forEach((resource) => {
         addSubscriptionDeployedResource(clusterId, resource, links, nodes)
     })
+}
+
+const processMultiples = (resources) => {
+    if (resources.length > 5) {
+        const groupByKind = groupBy(resources, 'kind')
+        return Object.entries(groupByKind).map(([kind, _resources]) => {
+            return {
+                kind,
+                name: '',
+                namespace: '',
+                resources: _resources,
+                resourceCount: _resources.length,
+            }
+        })
+    }
+    return resources
 }
 
 // Route, Ingress, StatefulSet
@@ -276,7 +284,7 @@ const processServiceOwner = (clusterId, serviceOwners, links, nodes, relatedReso
                     }
                     break
             }
-        } else {
+        } else if (serviceOwners.length === 1) {
             servicesMap[`serviceOwner${inx}`] = node.id
         }
     })
@@ -306,10 +314,11 @@ const addSubscriptionDeployedResource = (parentId, resource, links, nodes) => {
               parentId,
               parentName: parentNode.name,
               parentType: parentNode.type,
+              parentSpecs: parentNode.specs,
           }
         : undefined
 
-    const { name, namespace, template } = resource
+    const { name, namespace, template, resources, resourceCount } = resource
     const kind = resource.kind.toLowerCase()
     const memberId = `member--deployed-resource--${parentId}--${namespace}--${name}--${kind}`
 
@@ -322,6 +331,8 @@ const addSubscriptionDeployedResource = (parentId, resource, links, nodes) => {
         specs: {
             parent: parentObject,
             template,
+            resources,
+            resourceCount,
         },
     }
 
@@ -351,12 +362,16 @@ export const createReplicaChild = (parentObject, template, links, nodes) => {
         if (template && template.related) {
             const relatedMap = keyBy(template.related, 'kind')
             if (relatedMap['replicaset']) {
-                return createChildNode(parentObject, type, links, nodes)
+                const pNode = createChildNode(parentObject, type, links, nodes)
+                return createChildNode(pNode, 'pod', links, nodes)
             } else if (relatedMap['pod']) {
                 return createChildNode(parentObject, 'pod', links, nodes)
             }
         } else {
-            return createChildNode(parentObject, type, links, nodes)
+            const pNode = createChildNode(parentObject, type, links, nodes)
+            if (type === 'replicaset') {
+                return createChildNode(pNode, 'pod', links, nodes)
+            }
         }
     }
 }
